Added high frequency noise filter see ref. http://goo.gl/pXb6q3    and http://goo.gl/rCm8ry
For this revision we also used a different configuration with the vertical
B-field concentrated in a cylindrical region at the centre of the
simulation box. Models for the 200G 500G configs with uniform vertical
field result in instabilities. Results indicate that these may be due to
boundary conditions for the magnetic field.

For this revision we've also updated the routines for matlab visualising
field contours and time-distance plots